extract requestInit from downloadAttachment
