+++ ts/features/messages/saga/handleDownloadAttachment.ts
@@ -26,15 +26,8 @@ import {
   trackThirdPartyMessageAttachmentDownloadFailed,
   trackThirdPartyMessageAttachmentUnavailable
 } from "../analytics";
-import {
-  lollipopKeyTagSelector,
-  lollipopPublicKeySelector
-} from "../../lollipop/store/reducers/lollipop";
-import { generateKeyInfo } from "../../lollipop/saga";
-import { LollipopConfig } from "../../lollipop";
-import { lollipopRequestInit } from "../../lollipop/utils/fetch";
-import { isTestEnv } from "../../../utils/environment";
 import { getHeaderByKey } from "../utils/strings";
+import { handleRequestInit } from "./handleRequestInit";
 
 export const AttachmentsDirectoryPath =
   RNFS.CachesDirectoryPath + "/attachments";
@@ -94,7 +87,7 @@ export function* downloadAttachmentWorker(
       });
 
       const { method, attachmentFullUrl, headers } = yield* call(
-        generateReactNativeBlobUtilsFetchParameters,
+        handleRequestInit,
         attachment,
         bearerToken,
         uuid()
@@ -157,7 +150,7 @@ export function* downloadAttachmentWorker(
  * @param bearerToken
  * @param action
  */
-export function* handleDownloadAttachmentSaga(
+export function* handleDownloadAttachment(
   bearerToken: SessionToken,
   action: ActionType<typeof downloadAttachment.request>
 ) {
@@ -178,65 +171,3 @@ export function* handleDownloadAttachmentSaga(
     yield* put(downloadAttachment.cancel(action.payload));
   }
 }
-
-type HeaderType = { [key: string]: string };
-
-function* generateReactNativeBlobUtilsFetchParameters(
-  attachment: UIAttachment,
-  bearerToken: string,
-  nonce: string
-) {
-  const lollopopConfig: LollipopConfig = {
-    nonce
-  };
-
-  const keyTag = yield* select(lollipopKeyTagSelector);
-  const publicKey = yield* select(lollipopPublicKeySelector);
-  const keyInfo = yield* call(generateKeyInfo, keyTag, publicKey);
-
-  const attachmentFullUrl = attachment.resourceUrl.href;
-
-  const requestInit = {
-    headers: {
-      Authorization: `Bearer ${bearerToken}`
-    } as HeaderType,
-    method: "GET" as const
-  };
-
-  try {
-    const lollipopInit = yield* call(
-      lollipopRequestInit,
-      lollopopConfig,
-      keyInfo,
-      attachmentFullUrl,
-      requestInit
-    );
-    return reactNativeBlobUtilsFetchParametersFactory(
-      requestInit.method,
-      attachmentFullUrl,
-      (lollipopInit.headers ?? requestInit.headers) as HeaderType
-    );
-  } catch {
-    // We are not interested in doing anything here, since the
-    // later http call will be refused by the lollipop consumer
-  }
-
-  return reactNativeBlobUtilsFetchParametersFactory(
-    requestInit.method,
-    attachmentFullUrl,
-    requestInit.headers
-  );
-}
-
-const reactNativeBlobUtilsFetchParametersFactory = (
-  method: "GET",
-  attachmentFullUrl: string,
-  headers: HeaderType
-) => ({ method, attachmentFullUrl, headers });
-
-export const testableData = isTestEnv
-  ? {
-      generateReactNativeBlobUtilsFetchParameters,
-      reactNativeBlobUtilsFetchParametersFactory
-    }
-  : undefined;
